// src/types/supabase.ts
export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '12.2.3 (519615d)'
  }
  public: {
// ...
    Views: {
      active_items: {
        Row: {
          category: string | null
          city: string | null
          country: string | null
          created_at: string | null
          deleted_at: string | null
          description: string | null
          email: string | null
          expire_interval: unknown | null
          explicit_expire_at: string | null
          full_name: string | null
          id: string | null
          images: string[] | null
          is_have_whatsapp: boolean | null
          location: unknown | null
          phone_number: string | null
          postal_code: string | null
          published_at: string | null
          status: string | null
          status_updated_at: string | null
          street_name: string | null
          street_number: string | null
          title: string | null
          updated_at: string | null
          user_id: string | null
        }
        Insert: {
          category?: string | null
          city?: string | null
          country?: string | null
          created_at?: string | null
          deleted_at?: string | null
          description?: string | null
          email?: string | null
          expire_interval?: unknown | null
          explicit_expire_at?: string | null
          full_name?: string | null
          id?: string | null
          images?: string[] | null
          is_have_whatsapp?: boolean | null
          location?: unknown | null
          phone_number?: string | null
          postal_code?: string | null
          published_at?: string | null
          status?: string | null
          status_updated_at?: string | null
          street_name?: string | null
          street_number?: string | null
          title?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Update: {
          category?: string | null
          city?: string | null
          country?: string | null
          created_at?: string | null
          deleted_at?: string | null
          description?: string | null
          email?: string | null
          expire_interval?: unknown | null
          explicit_expire_at?: string | null
          full_name?: string | null
          id?: string | null
          images?: string[] | null
          is_have_whatsapp?: boolean | null
          location?: unknown | null
          phone_number?: string | null
          postal_code?: string | null
          published_at?: string | null
          status?: string | null
          status_updated_at?: string | null
          street_name?: string | null
          street_number?: string | null
          title?: string | null
          updated_at?: string | null
          user_id?: string | null
        }
        Relationships: []
      }
      geography_columns: {
        Row: {
          coord_dimension: number | null
          f_geography_column: unknown | null
          f_table_catalog: unknown | null
          f_table_name: unknown | null
          f_table_schema: unknown | null
          srid: number | null
          type: string | null
        }
        Relationships: []
      }
      geometry_columns: {
        Row: {
          coord_dimension: number | null
          f_geometry_column: unknown | null
          f_table_catalog: string | null
          f_table_name: unknown | null
          f_table_schema: unknown | null
          srid: number | null
          type: string | null
        }
        Insert: {
          coord_dimension?: number | null
          f_geometry_column?: unknown | null
          f_table_catalog?: string | null
          f_table_name?: unknown | null
          f_table_schema?: unknown | null
          srid?: number | null
          type?: string | null
        }
        Update: {
          coord_dimension?: number | null
          f_geometry_column?: unknown | null
          f_table_catalog?: string | null
          f_table_name?: unknown | null
          f_table_schema?: unknown | null
          srid?: number | null
          type?: string | null
        }
        Relationships: []
      }
      last_messages_per_conversation: {
        Row: {
          content: string | null
          conversation_id: string | null
          created_at: string | null
          id: string | null
          sender_id: string | null
        }
        Relationships: [
          {
            foreignKeyName: 'fk_conversation'
            columns: ['conversation_id']
            isOneToOne: false
            referencedRelation: 'conversations'
            referencedColumns: ['id']
          },
        ]
      }
    }
// ...
  }
}